+++ src/runner/bin.ts
@@ -1,4 +1,8 @@
 import index from './'
+import debug from '../debug'
+
+// Debug
+debug(`cwd: ${process.cwd()}`)
 
 // Run hook
 index(process.argv)